feat: rename circuit test scripts, update readme
