+++ README.md
@@ -4,13 +4,13 @@
 
 UniRep is a private and nonrepudiable repuation system. Users can receive positive and negative reputation from attesters, and voluntarily prove that they have at least certain amount of reputation without revealing the exact amount. Moreover, users cannot refuse to receive reputation from an attester.
 
-For more information, refer to the [documentation](https://vivi432.gitbook.io/unirep/)
+For more information, refer to the [documentation](https://unirep.gitbook.io/unirep/)
 
 ## Project Structure
 
 -   Composed by 4 packages:
     -   [`unirep`](./packages/unirep/) - main package.
-    -   [`contract`](./packages/contracts/) - Unirep smart contracts.
+    -   [`contracts`](./packages/contracts/) - Unirep smart contracts.
     -   [`circuits`](./packages/circuits/) - Circom libaries.
     -   [`crypto`](./packages/crypto) - Crypto utils.
