+++ packages/circuits/test/processAttestations.test.ts
