+++ packages/circuits/test/proveUserSignUp.test.ts
