+++ packages/circuits/test/sparseMerkleTree.test.ts
